road_searching/requirements.txt
torch==1.9.0
torchvision==0.10.0
torchsummary==1.5.1
torchmetrics==0.5.1
segmentation-models-pytorch==0.2.0
Pillow==8.3.2
matplotlib==3.4.3
numpy==1.21.2
